+++ src/api/__mocks__/peers.js
@@ -1,0 +1,4 @@
+export async function fetchPeers(authHost, token) {
+  const data = ['lain.com', 'heaven.com']
+  return Promise.resolve({ data })
+}