+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Changed
+
+- Statuses count changes when an instance is selected and shows the amount of statuses from an originating instance
+- Add a dialog window with a confirmation when a remove button is clicked on the Settings page
+
+### Fixed
+
+- Send `true` and `false` as booleans if they are values of single selects on the Settings page
+
 ## [2.0.3] - 2020-04-29
 
 ### Added